+++ .github/workflows/wildcard-go-binary.yml
@@ -3,7 +3,6 @@ on:
     branches:
       - wildcard
     paths:
-      - "pkg/**/*"
       - "internal/**/*"
       - "go.mod"
       - "main.go"
